// gen/ts/src/device/device.ts
export interface Device {
  descriptor?: DeviceDescriptor | undefined;
  properties?:
    | { $case: "thermostat"; thermostat: ThermostatProperties }
    | { $case: "temperatureSensor"; temperatureSensor: TemperatureSensorProperties }
    | { $case: "humiditySensor"; humiditySensor: HumiditySensorProperties }
    | { $case: "occupancySensor"; occupancySensor: OccupancySensorProperties }
    | { $case: "contactSensor"; contactSensor: ContactSensorProperties }
    | { $case: "windowCovering"; windowCovering: WindowCoveringProperties }
    | { $case: "doorLock"; doorLock: DoorLockProperties }
    | { $case: "onOffLight"; onOffLight: OnOffLightProperties }
    | { $case: "dimmableLight"; dimmableLight: DimmableLightProperties }
    | { $case: "colorTempLight"; colorTempLight: ColorTemperatureLightProperties }
    | { $case: "energySensor"; energySensor: EnergySensorProperties }
    | { $case: "pressureSensor"; pressureSensor: PressureSensorProperties }
    | //
    /** Vendor extensions (fields 50-99) */
    { $case: "homematic"; homematic: HomematicVendorExtension }
    | undefined;
  lastSeen?: Date | undefined;
  lastUpdated?: Date | undefined;
}